Move MoodleRepository to repository package
